mengubah response, menambahkan relationship

+++ 4-api/api/Controllers/EmployeesController.cs
@@ -1,3 +1,4 @@
+using System.Net;
 using api.Model;
 using api.Repositories;
 using Microsoft.AspNetCore.Mvc;
@@ -18,9 +19,9 @@ namespace api.Controllers
         public ActionResult Get() {
             var data = repository.Get(); // mengambil fungsi dari repositori
             if(data != null) {
-                return Ok(data);
+                return Ok(new { status = HttpStatusCode.OK, message = "Data berhasil ditemukan.", data = data });
             } else {
-                return NotFound("Data tidak tersedia.");
+                return NotFound(new { status = HttpStatusCode.NotFound, message = "Data tidak ditemukan."});
             }
         }
 
@@ -28,54 +29,46 @@ namespace api.Controllers
         public virtual ActionResult Get(string NIK) {
             var data = repository.Get(NIK);
             if(data != null) {
-                return Ok(data);
+                return Ok(new { status = HttpStatusCode.OK, message = "Data berhasil ditemukan.", data = data });
             } else {
-                return NotFound("Data tidak tersedia.");
+                return NotFound(new { status = HttpStatusCode.NotFound, message = "Data tidak ditemukan."});
             }
         }
 
         [HttpPost]
         public virtual ActionResult Insert(Employee employee) {
             if(repository.CheckEmailUnique(employee.Email)==true) {
-                return BadRequest("Email tidak boleh duplikat.");
+                return BadRequest(new { status = HttpStatusCode.BadRequest, message = "Email tidak boleh duplikat." });
             } else if(repository.CheckPhoneUnique(employee.Phone)==true) {
-                return BadRequest("Phone tidak boleh duplikat.");
+                return BadRequest(new { status = HttpStatusCode.BadRequest, message = "Phone tidak boleh duplikat." });
             }
             
-            var result = repository.Insert(employee);
-            if(result > 0) {
-                return Ok("Data berhasil ditambahkan.");
-            } else {
-                return Problem("Data tidak berhasil ditambahkan.");
-            }
+            repository.Insert(employee);
+            return Ok(new { status = HttpStatusCode.OK, message = "Data berhasil ditambahkan." });
         }
 
         [HttpDelete("{NIK}")]
         public virtual ActionResult Delete(string NIK) {
-            var result = repository.Delete(NIK);
-            if(result > 0) {
-                return Ok("Data berhasil dihapus.");
-            } else {
-                return NotFound("Data tidak tersedia.");
+            if(repository.CheckNIKExist(NIK)==false) {
+                return NotFound(new { status = HttpStatusCode.NotFound, message = "NIK tidak ditemukan." });
             }
+
+            repository.Delete(NIK);
+            return Ok(new { status = HttpStatusCode.OK, message = "Data berhasil dihapus." });
         }
 
         [HttpPut]
         public virtual ActionResult Update(Employee employee) {
             if(repository.CheckNIKExist(employee.NIK)==false) {
-                return NotFound("NIK tidak ditemukan.");
+                return NotFound(new { status = HttpStatusCode.NotFound, message = "NIK tidak ditemukan." });
             } else if(repository.CheckEmailUnique(employee.Email)==true) {
-                return BadRequest("Email tidak boleh duplikat.");
+                return BadRequest(new { status = HttpStatusCode.BadRequest, message = "Email telah digunakan."});
             } else if(repository.CheckPhoneUnique(employee.Phone)==true) {
-                return BadRequest("Phone tidak boleh duplikat.");
+                return BadRequest(new { status = HttpStatusCode.BadRequest, message = "Phone telah digunakan."});
             }
 
-            var result = repository.Update(employee);
-            if(result > 0) {
-                return Ok("Data berhasil diubah.");
-            } else {
-                return Problem("Data tidak berhasil diubah.");
-            }
+            repository.Update(employee);
+            return Ok(new { status = HttpStatusCode.OK, message = "Data berhasil diubah." });
         }
     }
 }
